refactor: Restructure table grid navigation utils
